DropInFixStoredPaymentMethodsOnly: remove PaymentMethodsListDelegate and add all the logic back to PaymentMethodsListViewModel
