+++ drop-in/src/main/java/com/adyen/checkout/dropin/ui/stored/StoredUtils.kt
@@ -36,3 +36,10 @@ internal fun StoredPaymentMethod.mapStoredModel(
         )
     }
 }
+
+internal fun StoredPaymentMethod.isStoredPaymentSupported(): Boolean {
+    return !this.type.isNullOrEmpty() &&
+        !this.id.isNullOrEmpty() &&
+        PaymentMethodTypes.SUPPORTED_PAYMENT_METHODS.contains(this.type) &&
+        this.isEcommerce
+}
